new weight function + modifications likend to the nmanual

+++ src/common/aka_error.hh
@@ -150,7 +150,7 @@ namespace debug {
         gettimeofday(&time, NULL);
         double  timestamp = time.tv_sec*1e6 + time.tv_usec; /*in us*/
         *(cout) << parallel_context
-                << "{" << timestamp << "} "
+                << "{" << (unsigned int)timestamp << "} "
                 << prefix << info << " "
                 << AKANTU_LOCATION
                 << std::endl;
